[#177] Refactor: Extract class, Move Method: Created a new class for managing the case location tracking expires so that we can do things when this happen.
Also extracted class and moved methods to LocationTrackingExpiredAlarmManager for this new receiver.
