+++ app/src/main/java/com/example/bigowlapp/utils/PeriodicLocationCheckAlarmManager.java
@@ -21,10 +21,9 @@ import com.google.android.gms.location.LocationServices;
  * Create alarm that will force a location check periodically
  */
 public class PeriodicLocationCheckAlarmManager {
-
     private static final int REQUEST_CODE = 5;
-    private static final int REQUEST_CODE_CANCEL = 6;
     private static final int NO_FLAG = 0;
+
     private final Context context;
     private final AlarmManager alarmManager;
 
@@ -36,10 +35,6 @@ public class PeriodicLocationCheckAlarmManager {
     public void setAlarm(long repeatIntervalMillis, long cancelTimeMillis) {
         alarmManager.setRepeating(AlarmManager.ELAPSED_REALTIME, SystemClock.elapsedRealtime(),
                 repeatIntervalMillis, getPendingIntent());
-
-        alarmManager.set(AlarmManager.ELAPSED_REALTIME,
-                SystemClock.elapsedRealtime() + cancelTimeMillis,
-                getCancelPendingIntent());
     }
 
     public void cancelPeriodicLocationCheck() {
@@ -51,11 +46,6 @@ public class PeriodicLocationCheckAlarmManager {
         return PendingIntent.getBroadcast(context, REQUEST_CODE, intent, NO_FLAG);
     }
 
-    private PendingIntent getCancelPendingIntent() {
-        Intent intent = new Intent(context, CancelPeriodicLocationCheckAlarmReceiver.class);
-        return PendingIntent.getBroadcast(context, REQUEST_CODE_CANCEL, intent, NO_FLAG);
-    }
-
     public static class PeriodicLocationCheckAlarmReceiver extends BroadcastReceiver {
         @Override
         public void onReceive(Context context, Intent intent) {
@@ -64,14 +54,6 @@ public class PeriodicLocationCheckAlarmManager {
                 return;
             }
             locationClient.getCurrentLocation(LocationRequest.PRIORITY_HIGH_ACCURACY, null);
-        }
-    }
-
-    public static class CancelPeriodicLocationCheckAlarmReceiver extends BroadcastReceiver {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            PeriodicLocationCheckAlarmManager locationCheckAlarmManager = new PeriodicLocationCheckAlarmManager(context);
-            locationCheckAlarmManager.cancelPeriodicLocationCheck();
         }
     }
 }
